+++ tests/integration/test_connector_field_defaults.rs
@@ -3,7 +3,6 @@
 use super::helpers::{data_parsing_fails, read_data_from_reader};
 
 use std::collections::HashMap;
-use std::path::PathBuf;
 
 use pathway_engine::connectors::data_format::{
     DsvParser, DsvSettings, InnerSchemaField, JsonLinesParser, ParsedEvent,
@@ -27,7 +26,7 @@ fn test_dsv_with_default_end_of_line() -> eyre::Result<()> {
     );
 
     let reader = CsvFilesystemReader::new(
-        PathBuf::from("tests/data/dsv_with_skips.txt"),
+        "tests/data/dsv_with_skips.txt",
         builder,
         ConnectorMode::Static,
         None,
@@ -81,7 +80,7 @@ fn test_dsv_with_default_middle_of_line() -> eyre::Result<()> {
     );
 
     let reader = CsvFilesystemReader::new(
-        PathBuf::from("tests/data/dsv_with_skips2.txt"),
+        "tests/data/dsv_with_skips2.txt",
         builder,
         ConnectorMode::Static,
         None,
@@ -131,7 +130,7 @@ fn test_dsv_fails_without_default() -> eyre::Result<()> {
     schema.insert("number".to_string(), InnerSchemaField::new(Type::Int, None));
 
     let reader = CsvFilesystemReader::new(
-        PathBuf::from("tests/data/dsv_with_skips.txt"),
+        "tests/data/dsv_with_skips.txt",
         builder,
         ConnectorMode::Static,
         None,
@@ -164,7 +163,7 @@ fn test_dsv_with_default_nullable() -> eyre::Result<()> {
     );
 
     let reader = CsvFilesystemReader::new(
-        PathBuf::from("tests/data/dsv_with_skips.txt"),
+        "tests/data/dsv_with_skips.txt",
         builder,
         ConnectorMode::Static,
         None,
@@ -208,7 +207,7 @@ fn test_dsv_with_default_nullable() -> eyre::Result<()> {
 #[test]
 fn test_jsonlines_fails_without_default() -> eyre::Result<()> {
     let reader = FilesystemReader::new(
-        PathBuf::from("tests/data/jsonlines.txt"),
+        "tests/data/jsonlines.txt",
         ConnectorMode::Static,
         None,
         ReadMethod::ByLine,
@@ -237,7 +236,7 @@ fn test_jsonlines_with_default() -> eyre::Result<()> {
     );
 
     let reader = FilesystemReader::new(
-        PathBuf::from("tests/data/jsonlines_with_skips.txt"),
+        "tests/data/jsonlines_with_skips.txt",
         ConnectorMode::Static,
         None,
         ReadMethod::ByLine,
@@ -290,7 +289,7 @@ fn test_jsonlines_with_default_at_jsonpath() -> eyre::Result<()> {
     );
 
     let reader = FilesystemReader::new(
-        PathBuf::from("tests/data/jsonlines_with_skips.txt"),
+        "tests/data/jsonlines_with_skips.txt",
         ConnectorMode::Static,
         None,
         ReadMethod::ByLine,
@@ -337,7 +336,7 @@ fn test_jsonlines_explicit_null_not_overridden() -> eyre::Result<()> {
     );
 
     let reader = FilesystemReader::new(
-        PathBuf::from("tests/data/jsonlines_with_skips_and_nulls.txt"),
+        "tests/data/jsonlines_with_skips_and_nulls.txt",
         ConnectorMode::Static,
         None,
         ReadMethod::ByLine,